Renamed modules which did not comply with PEP8 naming conventions
