+++ tests/test_dict_reader.py
